requirements.txt
et-xmlfile==1.1.0
excel2json==1.0.1
openpyxl==3.1.3
pandas==2.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
tzdata==2024.1
xlrd==2.0.1
asgiref==3.8.1      
django==5.0.6
django-jazzmin==3.0.0
django-cors-headers==4.4.0
# pip==22.2.2
setuptools==63.2.0
sqlparse==0.5.0
typing_extensions==4.12.2
djangorestframework==3.14.0
xhtml2pdf==0.2.16
weasyprint==63.0